refactor(web): split features/feedback into -layout/feedback and components/feedback
MON-847. Audit revealed use-survey-modal hook has global consumers (__root, blocks/early-access-banner, -context-panel/context-panel-rail), so hook stays in apps/web/src/hooks. Survey modal primitives (modal content + question) move alongside global hook in apps/web/src/components/feedback. Dashboard-only bits (auto-bug-reporter, monthly-satisfaction-survey + their tracker hook + constants) move to -layout/feedback.
