+++ apps/web/src/routes/_authenticated/$slug/$teamSlug/-layout/feedback/auto-bug-reporter.tsx
@@ -1,7 +1,7 @@
 import { POSTHOG_SURVEYS } from "@core/posthog/config";
 import { useSurveyModal } from "@/hooks/use-survey-modal";
 import { useEffect } from "react";
-import { useApiErrorTracker } from "../hooks/use-api-error-tracker";
+import { useApiErrorTracker } from "./use-api-error-tracker";
 
 export function AutoBugReporter() {
    const { shouldShowBugReport, dismiss } = useApiErrorTracker();
